fix(desktop): route multi-workflow sessions by step ownership (phase 10)
activateWorkflowAgent, sendTurn turn-routing, and the step-completion
cursor update all hardcoded session.workflowIds[0]. With more than one
workflow attached, clicking the lit step of workflow #2+ would silently
produce an empty kickoff (template lookup missed, promptPrefix='',
kickoff.length===0, sendTurn never fires).

Route by step ownership across all attached workflows. Also collapse
the duplicate WorkflowNextStepCta/index.test.ts (phase-2 hygiene miss)
into index.test.tsx with the comprehensive pickNextWorkflowStep suite.
